requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Werkzeug==3.0.1
Transformers==4.38.2
Pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
opencv-python==4.9.0.80
face-recognition==1.3.0
python-dotenv==1.0.1
tensorflow-macos==2.16.1
scikit-learn==1.4.0